+++ s2-site/examples/react-component/sheet/demo/table.tsx
@@ -3,7 +3,9 @@ import ReactDOM from 'react-dom';
 import { SheetComponent } from '@antv/s2-react';
 import '@antv/s2-react/dist/style.min.css';
 
-fetch('../data/basic.json')
+fetch(
+  'https://gw.alipayobjects.com/os/bmw-prod/d62448ea-1f58-4498-8f76-b025dd53e570.json',
+)
   .then((res) => res.json())
   .then((res) => {
     const s2Options = {